requirements.txt
asgiref==3.9.2
click==8.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.6
django-modeltranslation==0.19.17
djangorestframework==3.16.1
drf-yasg==1.21.11
gunicorn==23.0.0
h11==0.16.0
inflection==0.5.1
packaging==25.0
polib==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pillow==11.3.0
psycopg2-binary
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
uvicorn==0.37.0
waitress==3.0.2
whitenoise==6.6.0
